Added conda environment output to WISER.spec and cleaned \n up submodule collection code in WISER.spec
