+++ WISER.spec
@@ -4,10 +4,14 @@ import sys ; sys.setrecursionlimit(sys.getrecursionlimit() * 5)
 from PyInstaller.utils.hooks import collect_all, collect_submodules
 from PyInstaller.building.build_main import Analysis, PYZ, EXE, COLLECT
 
-from pathlib import Path
-import json, subprocess, shlex, os, sys
-import importlib.metadata as imeta
-from packaging.version import Version, InvalidVersion
+import subprocess
+
+# Create the dependency list file
+result = subprocess.run(
+    [sys.executable, "src/devtools/write_dependencies.py"],
+    capture_output=True,
+    text=True,
+)
 
 datas = [('src\\wiser\\bandmath\\bandmath.lark', 'wiser\\bandmath')]
 binaries = [('C:\\Users\\jgarc\\anaconda3\\envs\\wiser-source\\Library\\plugins\\platforms', 'platforms'), ('C:\\Users\\jgarc\\anaconda3\\envs\\wiser-source\\Library\\plugins\\iconengines', 'iconengines'), ('C:\\Users\\jgarc\\anaconda3\\envs\\wiser-source\\Library\\lib\\gdalplugins\\gdal_FITS.dll', 'gdalplugins'), ('C:\\Users\\jgarc\\anaconda3\\envs\\wiser-source\\Library\\lib\\gdalplugins\\gdal_netCDF.dll', 'gdalplugins'), ('C:\\Users\\jgarc\\anaconda3\\envs\\wiser-source\\Library\\lib\\gdalplugins\\gdal_HDF4.dll', 'gdalplugins'), ('C:\\Users\\jgarc\\anaconda3\\envs\\wiser-source\\Library\\lib\\gdalplugins\\gdal_HDF5.dll', 'gdalplugins')]
@@ -30,9 +34,6 @@ temp_a = Analysis(
     optimize=0,
 )
 
-print(f"------------------------108j310u4nu10831904141vn41n4891y----------------------------")
-
-print(f"temp_a.pure: {temp_a.pure}")
 # BUILD UP NEW hiddenimports by collecting submodules for every top-level package
 top_modules = { entry[0].split('.', 1)[0] for entry in temp_a.pure }
 
@@ -48,20 +49,17 @@ for pkg in sorted(top_modules):
         continue
     hiddenimports.extend(collect_submodules(pkg))
 
-print(f"top_modules: {top_modules}")
-
-# remove duplicates while preserving order
+# Remove duplicates while preserving order
 _seen = set()
 _hidden = []
 for m in hiddenimports:
     if m not in _seen:
         _seen.add(m)
         _hidden.append(m)
-print(f"hidden: {_hidden}")
 hiddenimports = _hidden
 
-print(f"------------------------108j310u4nu10831904141vn41n4891y----------------------------")
-# 4) SECOND PASS: rebuild Analysis with the full hiddenimports list
+
+# SECOND PASS: rebuild Analysis with the full hiddenimports list
 a = Analysis(
     ['src\\wiser\\__main__.py'],
     pathex=[],
@@ -76,7 +74,6 @@ a = Analysis(
     optimize=0,
 )
 
-# 5) the rest of your spec unchanged
 pyz = PYZ(a.pure, a.zipped_data)
 
 exe = EXE(
@@ -107,4 +104,3 @@ coll = COLLECT(
     upx_exclude=[],
     name='WISER',
 )
-print(f"Done!")
