--- env-test.yml
@@ -1,35 +1,0 @@
-name: wiser-source-2
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - astropy
-  - bottleneck
-  - cfitsio
-  - conda-forge::blas[build=openblas]
-  - dask
-  - gdal
-  - ipykernel
-  - jax
-  - libgdal-fits
-  - libgdal-netcdf
-  - matplotlib
-  - netcdf4
-  - numba=0.60.0
-  - numcodecs
-  - numpy=1.23.5
-  - openjpeg
-  - pdr
-  - pyinstaller
-  - pyside2
-  - python=3.10
-  - scipy=1.9.2
-  - xarray
-  - zlib
-  - libgdal-jp2openjpeg
-  - rasterio
-  - glymur
-  - pyproj
-  - pytest=8.3.4
-  - pluggy=1.5.0
-prefix: C:\Users\jgarc\anaconda3\envs\wiser-source-2
