+++ src/devtools/README.md
@@ -1,0 +1,3 @@
+# Development Tools
+
+This directory is used for developer tooling. Currently, its use is to house files needed for collecting all of the dependencies of the WISER build. We need these dependencies listed out so that we can tell the user what packages they can and can't use.
