--- wiser_dependencies_1.2b1_windows.json
@@ -1,185 +1,0 @@
-{
-  "channels": [
-    "conda-forge",
-    "defaults"
-  ],
-  "dependencies": [
-    "_libavif_api=1.3.0=h57928b3_0",
-    "_openmp_mutex=4.5=2_gnu",
-    "altgraph=0.17.4=pyhd8ed1ab_1",
-    "aom=3.9.1=he0c23c2_0",
-    "astropy=6.0.1=py39hd88c2e4_0",
-    "astropy-iers-data=0.2025.6.9.0.39.3=pyhd8ed1ab_0",
-    "blas=2.132=openblas",
-    "blas-devel=3.9.0=32_hc0f8095_openblas",
-    "blosc=1.21.6=hfd34d9b_1",
-    "brotli=1.1.0=h2466b09_3",
-    "brotli-bin=1.1.0=h2466b09_3",
-    "bzip2=1.0.8=h2466b09_7",
-    "ca-certificates=2025.6.15=h4c7d964_0",
-    "certifi=2025.6.15=pyhd8ed1ab_0",
-    "cftime=1.6.4=py39h4b0a98a_1",
-    "contourpy=1.3.0=py39h2b77a98_2",
-    "cycler=0.12.1=pyhd8ed1ab_1",
-    "cytoolz=1.0.1=py39ha55e580_0",
-    "dav1d=1.2.1=hcfcfb64_0",
-    "dustgoggles=0.7.6=pyhd8ed1ab_0",
-    "fonttools=4.58.5=py39h5769e4c_0",
-    "freetype=2.13.3=h57928b3_1",
-    "freexl=2.0.0=hf297d47_2",
-    "future=1.0.0=pyhd8ed1ab_2",
-    "gdal=3.10.2=py39hbed5357_5",
-    "geos=3.13.1=h9ea8674_0",
-    "geotiff=1.7.4=h887f4e7_0",
-    "glib=2.84.1=h3d4babf_0",
-    "glib-tools=2.84.1=h4394cf3_0",
-    "gst-plugins-base=1.24.11=h3fe0a9e_0",
-    "gstreamer=1.24.11=h233a61a_0",
-    "hdf4=4.2.15=h5557f11_7",
-    "hdf5=1.14.6=nompi_hd5d9e70_101",
-    "icu=75.1=he0c23c2_0",
-    "importlib-metadata=8.7.0=pyhe01879c_1",
-    "importlib-resources=6.5.2=pyhd8ed1ab_0",
-    "importlib_resources=6.5.2=pyhd8ed1ab_0",
-    "kiwisolver=1.4.7=py39h2b77a98_0",
-    "krb5=1.21.3=hdf4eb48_0",
-    "lcms2=2.17=hbcf6048_0",
-    "lerc=4.0.0=h6470a55_1",
-    "libaec=1.1.4=h20038f6_0",
-    "libarchive=3.7.7=h5343c79_4",
-    "libavif16=1.3.0=hf2698fe_0",
-    "libblas=3.9.0=32_h11dc60a_openblas",
-    "libbrotlicommon=1.1.0=h2466b09_3",
-    "libbrotlidec=1.1.0=h2466b09_3",
-    "libbrotlienc=1.1.0=h2466b09_3",
-    "libcblas=3.9.0=32_h9bd4c3b_openblas",
-    "libclang13=20.1.7=default_h6e92b77_0",
-    "libcurl=8.14.1=h88aaa65_0",
-    "libde265=1.0.15=h91493d7_0",
-    "libdeflate=1.23=h76ddb4d_0",
-    "libexpat=2.7.0=he0c23c2_0",
-    "libffi=3.4.6=h537db12_1",
-    "libfreetype=2.13.3=h57928b3_1",
-    "libfreetype6=2.13.3=h0b5ce68_1",
-    "libgcc=15.1.0=h1383e82_3",
-    "libgdal-core=3.10.2=h4d2c634_1",
-    "libglib=2.84.1=h7025463_0",
-    "libgomp=15.1.0=h1383e82_3",
-    "libheif=1.19.7=gpl_h2684147_100",
-    "libhwy=1.2.0=h1d1702c_0",
-    "libiconv=1.18=h135ad9c_1",
-    "libintl=0.22.5=h5728263_3",
-    "libintl-devel=0.22.5=h5728263_3",
-    "libjpeg-turbo=3.1.0=h2466b09_0",
-    "libjxl=0.11.1=ha161b08_2",
-    "libkml=1.3.0=h538826c_1021",
-    "liblapack=3.9.0=32_h2526c6b_openblas",
-    "liblapacke=3.9.0=32_h1d0e49f_openblas",
-    "liblzma=5.8.1=h2466b09_2",
-    "libnetcdf=4.9.2=nompi_he045f6b_117",
-    "libogg=1.3.5=h2466b09_1",
-    "libopenblas=0.3.30=pthreads_ha4fe6b2_0",
-    "libpng=1.6.50=h95bef1e_0",
-    "librttopo=1.1.0=hbfc9ebc_18",
-    "libspatialite=5.1.0=h208f9ff_13",
-    "libsqlite=3.50.2=hf5d6505_0",
-    "libssh2=1.11.1=h9aa295b_0",
-    "libtiff=4.7.0=h797046b_4",
-    "libvorbis=1.3.7=h0e60522_0",
-    "libwebp-base=1.5.0=h3b0e114_0",
-    "libwinpthread=12.0.0.r4.gg4f2fc60ca=h57928b3_9",
-    "libxcb=1.17.0=h0e4246c_0",
-    "libxml2=2.13.8=h442d1da_0",
-    "libxslt=1.1.39=h3df6e99_0",
-    "libzip=1.11.2=h3135430_0",
-    "libzlib=1.3.1=h2466b09_2",
-    "llvmlite=0.43.0=py39he94c479_1",
-    "lz4-c=1.10.0=h2466b09_1",
-    "lzo=2.10=hcfcfb64_1001",
-    "matplotlib=3.9.1=py39hcbf5309_1",
-    "matplotlib-base=3.9.1=py39h5376392_2",
-    "minizip=4.0.10=h9fa1bad_0",
-    "more-itertools=10.7.0=pyhd8ed1ab_0",
-    "multidict=6.6.3=py39h5769e4c_0",
-    "munkres=1.1.4=pyhd8ed1ab_1",
-    "muparser=2.3.5=he0c23c2_0",
-    "netcdf4=1.7.2=nompi_py39h38b60f6_102",
-    "numba=0.60.0=py39h5dcb127_0",
-    "numpy=1.23.5=py39hbccbffa_0",
-    "openblas=0.3.30=pthreads_h4a7f399_0",
-    "openjpeg=2.5.3=h4d64b90_0",
-    "openssl=3.5.1=h725018a_0",
-    "packaging=25.0=pyh29332c3_1",
-    "pandas=2.3.0=py39h2366fc2_0",
-    "pcre2=10.44=h99c9b8b_2",
-    "pdr=1.4.0=pyhd8ed1ab_0",
-    "pefile=2023.2.7=pyhd8ed1ab_0",
-    "pillow=11.3.0=py39hbad85af_0",
-    "pip=25.1.1=pyh8b19718_0",
-    "ply=3.11=pyhd8ed1ab_3",
-    "proj=9.5.1=h4f671f6_0",
-    "pthread-stubs=0.4=h0e40799_1002",
-    "pvl=1.3.2=pyhd8ed1ab_1",
-    "pyerfa=2.0.1.5=py39h4b0a98a_0",
-    "pyinstaller=6.14.1=py39h2e88fad_0",
-    "pyinstaller-hooks-contrib=2025.5=pyhd8ed1ab_0",
-    "pyparsing=3.2.3=pyhd8ed1ab_1",
-    "pyproj=3.6.1=py39h4afcdfd_10",
-    "pyqt=5.15.11=py39he15501f_1",
-    "pyqt5-sip=12.17.0=py39ha51f57c_1",
-    "pyside2=5.15.15=py39h0285922_0",
-    "python=3.9.23=h8c5b53a_0_cpython",
-    "python-dateutil=2.9.0.post0=pyhe01879c_2",
-    "python-tzdata=2025.2=pyhd8ed1ab_0",
-    "python_abi=3.9=7_cp39",
-    "pytz=2025.2=pyhd8ed1ab_0",
-    "pywin32=307=py39ha51f57c_3",
-    "pywin32-ctypes=0.2.3=py39hcbf5309_1",
-    "pyyaml=6.0.2=py39hf73967f_2",
-    "qhull=2020.2=hc790b64_5",
-    "qt-main=5.15.15=h9151539_4",
-    "rav1e=0.7.1=ha073cba_3",
-    "rms-vax=1.0.5=pyhd8ed1ab_1",
-    "scipy=1.13.1=py39h1a10956_0",
-    "setuptools=75.8.0=pyhff2d567_0",
-    "sip=6.10.0=py39ha51f57c_0",
-    "six=1.17.0=pyhd8ed1ab_0",
-    "snappy=1.2.1=h500f7fa_1",
-    "sqlite=3.50.2=hdb435a2_0",
-    "svt-av1=3.0.2=he0c23c2_0",
-    "tk=8.6.13=h2c6b04d_2",
-    "toml=0.10.2=pyhd8ed1ab_1",
-    "tomli=2.2.1=pyhd8ed1ab_1",
-    "toolz=1.0.0=pyhd8ed1ab_1",
-    "tornado=6.5.1=py39ha55e580_0",
-    "typing-extensions=4.14.0=h32cad80_0",
-    "typing_extensions=4.14.0=pyhe01879c_0",
-    "tzdata=2025b=h78e105d_0",
-    "ucrt=10.0.22621.0=h57928b3_1",
-    "unicodedata2=16.0.0=py39ha55e580_0",
-    "uriparser=0.9.8=h5a68840_0",
-    "vc=14.3=h41ae7f8_26",
-    "vc14_runtime=14.44.35208=h818238b_26",
-    "vs2015_runtime=14.44.35208=h38c0c73_26",
-    "wheel=0.45.1=pyhd8ed1ab_1",
-    "x265=3.5=h2d74725_3",
-    "xerces-c=3.2.5=he0c23c2_2",
-    "xorg-libxau=1.0.12=h0e40799_0",
-    "xorg-libxdmcp=1.1.5=h0e40799_0",
-    "yaml=0.2.5=h8ffe710_2",
-    "zipp=3.23.0=pyhd8ed1ab_0",
-    "zlib=1.3.1=h2466b09_2",
-    "zstd=1.5.7=hbeecb71_2",
-    {
-      "pip": [
-        "bugsnag==4.7.1",
-        "lark==1.2.2",
-        "opencv-python==4.5.3.56",
-        "psutil==7.0.0",
-        "webob==1.8.9"
-      ]
-    }
-  ],
-  "name": "wiser-strict-env",
-  "prefix": "C:\\Users\\jgarc\\anaconda3\\envs\\wiser-strict-env"
-}
