--- wiser_frozen_requirements1.txt
@@ -1,29 +1,0 @@
-astropy==6.1.7
-bokeh==2.4.3
-Bottleneck==1.5.0
-Brotli==1.1.0
-bugsnag==4.7.1
-cffi==1.17.1
-dask==2023.3.0
-distributed==2023.3.0
-GDAL==3.10.3
-h2==4.2.0
-importlib_metadata==8.6.1
-importlib_resources==6.5.2
-lark==1.2.2
-lz4==4.4.4
-netCDF4==1.7.2
-numba==0.60.0
-numcodecs==0.13.1
-numpy==1.23.5
-pdr==1.3.0
-pickleshare==0.7.5
-pluggy==1.5.0
-ply==3.11
-pvl==1.3.2
-pyproj==3.7.1
-PySide2==5.15.15
-pytest==8.3.4
-scipy==1.9.2
-zstandard==0.23.0
-opencv-python==4.5.3.56
